Move `methodReturnNode` from AstCreator to NodeBuilders
