+++ joern-cli/frontends/javasrc2cpg/src/main/scala/io/joern/javasrc2cpg/passes/AstCreator.scala
@@ -94,6 +94,7 @@ import io.joern.x2cpg.utils.NodeBuilders.{
   callNode,
   fieldIdentifierNode,
   identifierNode,
+  methodReturnNode,
   modifierNode,
   operatorCallNode
 }
